+++ src/neural_network/layer.rs
@@ -1,17 +1,18 @@
+pub mod conv2d;
 pub mod dense;
 pub mod lstm;
 pub mod simple_rnn;
 
+pub use conv2d::*;
 pub use dense::*;
-
-pub use simple_rnn::*;
-
 pub use lstm::*;
+pub use simple_rnn::*;
 
 pub enum LayerWeight<'a> {
     Dense(DenseLayerWeight<'a>),
     SimpleRNN(SimpleRNNLayerWeight<'a>),
     LSTM(LSTMLayerWeight<'a>),
+    Conv2D(Conv2DLayerWeight<'a>),
     Empty,
 }
 
@@ -37,4 +38,9 @@ pub struct LSTMLayerWeight<'a> {
     pub forget: LSTMGateWeight<'a>,
     pub cell: LSTMGateWeight<'a>,
     pub output: LSTMGateWeight<'a>,
+}
+
+pub struct Conv2DLayerWeight<'a> {
+    pub weight: &'a ndarray::Array4<f32>,
+    pub bias: &'a ndarray::Array2<f32>,
 }